Add files via upload
In the controller file, I added code to get the tasks to move from one column to another with a drag and drop, but the tasks are just being duplicated in the same column so I am not sure what I forgot. I set up a check to see what column it is in, but I don't know what's wrong. If someone wants to look and see if they can see what the problem is they can.